# Implement SKIP LOCKED during action claiming