--- a/.github/workflows/pr-unit-tests.yml
+++ b/.github/workflows/pr-unit-tests.yml
@@ -10,10 +10,10 @@ jobs:
       fail-fast: false
       matrix:
         # We test against the earliest and latest PHP versions for each major supported version.
-        php: [ '8.3' ]
-        wp: [ 'latest' ]
-        multisite: [ '0' ]
-        db: [ 'mariadb:10.6']
+        php: [ '7.1', '7.4', '8.0', '8.3' ]
+        wp: [ '6.5', '6.6', 'latest', 'nightly' ]
+        multisite: [ '0', '1' ]
+        db: [ 'mysql:5.6', 'mysql:8.1', 'mariadb:10.4', 'mariadb:10.6']
         exclude:
           # WordPress 6.6+ requires PHP 7.2+
           - php: 7.1