### Avoiding SQL deadlocks during action claiming

When running multiple action scheduler queues concurrently, many will start to fatal error due to SQL deadlocks. I have found that this always stems from the action claims query. This query ends up scanning many rows, and being an UPDATE, it ends up creating locks on the table/rows. This results in concurrent claims queries conflicting with each other and deadlocking. It also causes deadlocks during individual action status updates from other queues that are already processing actions. This leaves actions in unknown states and results in lots needing to be failed later on due to timeouts during the cleanup periods.

This PR prevents the deadlocks altogether. The magic here comes from the `SKIP LOCKED` SQL feature: https://dev.mysql.com/blog-archive/mysql-8-0-1-using-skip-locked-and-nowait-to-handle-hot-rows/. It's a perfect fit for the claims query, because we don't really care if it needs to skip over a few rows. All we want is for it to claim actions without conflict.

The `SKIP LOCKED` has to come from a SELECT query. So the new query here uses a SELECT sub-query (w/ `FOR UPDATE` to claim a lock) and then JOINs the results to accomplish first selecting the items and then updating them all within a single transaction.

I've done a lot of testing with this and ran into no deadlocks or failing actions. I've run hundreds of thousands of actions at 250 queue concurrency without any signs of trouble.

```
BEFORE (causes deadlocks):

UPDATE wp_actionscheduler_actions SET claim_id=4322, last_attempt_gmt='2025-01-15 03:24:52', last_attempt_local='2025-01-14 21:24:53' WHERE claim_id = 0 AND scheduled_date_gmt <= '2025-01-15 03:23:18' AND status='pending' ORDER BY priority ASC, attempts ASC, scheduled_date_gmt ASC, action_id ASC LIMIT 20;

AFTER (no deadlocks):

UPDATE wp_actionscheduler_actions t1 JOIN ( SELECT action_id from wp_actionscheduler_actions WHERE claim_id = 0 AND scheduled_date_gmt <= '2025-01-15 03:24:52' AND status='pending' ORDER BY priority ASC, attempts ASC, scheduled_date_gmt ASC, action_id ASC LIMIT 20 FOR UPDATE SKIP LOCKED ) t2 ON t1.action_id = t2.action_id SET claim_id=4322, last_attempt_gmt='2025-01-15 03:24:52', last_attempt_local='2025-01-14 21:24:53';
```

### Considerations

I'm not necessarily suggesting we merge this PR as is (leaving as a draft for now), but I wanted to show the solution here and what I've found. We'll need to test and discover database compatibility before shipping this as the new default behavior.

`SKIP LOCKED` shipped with MySQL 8.0.1 (2017-04-10) and MariaDB in 10.6.0 (2021-4-26), and works only for InnoDB tables (ignored otherwise). I've only done the testing in MySQL only, and I'm not sure if it's safely ignored on the older versions or if it would cause issues.

I do think this would definitely be a worthwhile change to get into the plugin, as I've seen it be a source of a lot of issues. Some possible rollout options:

A) Put it into a new method (`claim_actions_v2()`), and use the new method when a feature flag is present via some constant or filter.
B) Check the database version, and automatically use the v2 method when it's modern enough.