@prettyboymp 2025-04-07
I've done some testing against this and am happy with the results.  So I've made some additional changes to deal with compatibility:

- Claim process always use the `UPDATE .... JOIN (SELECT ... FOR UPDATE ... WHERE.... ORDER BY...)` format (without `SKIP LOCKED` if not supported) rather than the previous `UPDATE... WHERE... ORDER BY...`.   

  This alone should get rid of most deadlocks since the latter is non-deterministic because the db doesn't lock the rows until attempting to update them when another process may have already locked a subset of the rows.  By moving to a `SELECT .. FOR UPDATE` the locks are set during the select process and claimed in a known order.  This can still result in long running lock waits for claim queries, but they should at least not result in deadlocks.

- Use the `SKIP LOCKED` modifier on the the sub-select only on DB versions that support it.  

  This should further improve performance for DB versions that support it by preventing the claim process from stacking caused by multiple concurrent processes attempting to wait on locks to clear by just continuing with the next records.

## Testing Instructions

1. Setup a site running Action Scheduler - it can be installed as a separate plugin.
2. Install [AS Pressure Cooker](https://github.com/woocommerce/as-pressure-cooker) as a plugin.  This PR has some tweaks/fixes for PHP8 and improved details on the report: https://github.com/woocommerce/as-pressure-cooker/pull/6
3. Watch your site's error_log and slow query log (if available).
4. Run the AS Pressure Cooker to test the changes:
  ```
  wp action-scheduler pressure-cooker --starting_queue_runners=1 --finishing_queue_runners=30 --initial_past_due_actions=60000
  ```

Notes when testing:
* The pressure cooker will start with `--starting_queue_runners` and end with `--finishing_queue_runners`, allowing you to see where errors start occurring, if any.
* You may be limited in how many runners you can test with based on the memory and connection limits of your environment.  So it may be a good idea to start with a lower number to see where your environment breaks down and then run a full test just below that limit.
* `--initial_past_due_actions` determines how many actions are created prior to starting the test.  More will be generated during the test, however, that generation may not be able to keep up with the processing.  So it is a good idea to set this high enough to make sure the processing rate isn't brought down by there not being enough actions to be process.

## Testing Results

I ran the following tests locally with the below diff applied to better simulate how action scheduler runs normally through requests and the limitation of an assumed 30s request timeout and forcing the clean up queries to be run.

```
diff --git a/classes/WP_CLI/ActionScheduler_WPCLI_QueueRunner.php b/classes/WP_CLI/ActionScheduler_WPCLI_QueueRunner.php
index 08cb0cd..150b470 100644
--- a/classes/WP_CLI/ActionScheduler_WPCLI_QueueRunner.php
+++ b/classes/WP_CLI/ActionScheduler_WPCLI_QueueRunner.php
@@ -60,6 +60,9 @@ class ActionScheduler_WPCLI_QueueRunner extends ActionScheduler_Abstract_QueueRu
 	 * @throws \WP_CLI\ExitException When there are too many concurrent batches.
 	 */
 	public function setup( $batch_size, $hooks = array(), $group = '', $force = false ) {
+		if( $this->get_execution_time() >= 28 ) {
+			return 0;
+		}
 		$this->run_cleanup();
 		$this->add_hooks();
 
@@ -120,6 +123,10 @@ class ActionScheduler_WPCLI_QueueRunner extends ActionScheduler_Abstract_QueueRu
 
 			$this->process_action( $action_id, $context );
 			$this->progress_bar->tick();
+
+			if( $this->get_execution_time() >= 28 ) {
+				break;
+			}
 		}
 
 		$completed = $this->progress_bar->current();
```

### This PR - MySQL 8.0.4 - InnoDB Engine

```
wp action-scheduler pressure-cooker --starting_queue_runners=30 --finishing_queue_runners=30 --initial_past_due_actions=60000

Deleting existing actions and claims.
Generating test actions  100% [===================================================================================================] 1:34 / 1:35
Created 60000 past due actions.

Elapsed:              120s (100%)
Generated actions:    64400
Processed actions:    62180
Processing Rate:      518.2/s
Active queue runners: 28/30
Replacement rate      6/28
Database errors:      none detected

Elapsed:              120s (100%)
Generated actions:    64343
Processed actions:    52066
Processing Rate:      433.9/s
Active queue runners: 30/30
Replacement rate      0/30
Database errors:      none detected

Test completed.
```

### TRUNK - MySQL 8.0.4 - InnoDB Engine

```
wp action-scheduler pressure-cooker --starting_queue_runners=30 --finishing_queue_runners=30 --initial_past_due_actions=60000

Deleting existing actions and claims.
Generating test actions  100% [===================================================================================================] 1:29 / 1:29
Created 60000 past due actions.

Elapsed:              120s (100%)
Generated actions:    64400
Processed actions:    31006 (654 fails)
Processing Rate:      258.4/s
Active queue runners: 29/30
Replacement rate      0/29
Database errors:      766 (766 deadlocks)

Test completed.
```

### This PR - MySQL 8.0.4 - MyISAM Engine (SKIP LOCKED ignored)

This was really just run to verify no issues were caused on systems using MyISAM engines.

```
wp action-scheduler pressure-cooker --starting_queue_runners=30 --finishing_queue_runners=30 --initial_past_due_actions=60000

Deleting existing actions and claims.
Generating test actions  100% [===================================================================================================] 1:29 / 1:29
Created 60000 past due actions.

Elapsed:              120s (100%)
Generated actions:    64324
Processed actions:    16707
Processing Rate:      139.2/s
Active queue runners: 30/30
Replacement rate      0/30
Database errors:      none detected

Test completed.
```